+++ R/transition_matrix.R
@@ -3,3 +3,6 @@ transition_matrix <- function(x)
 
 observation_matrix <- function(x) 
   .translate_probabilities(x, field = "observation_prob", from = "states", to = "observations")
+
+reward_matrix <- function(x)
+  .translate_reward(x)